upd 4 hw

+++ controllers/contactsControllers.js
@@ -1,16 +1,17 @@
 import express from "express";
 
-import { Contact } from '../models/contact.js';
+import { Contact } from "../models/contact.js";
 import HttpError from "../helpers/HttpError.js";
 import {
   createContactSchema,
-    updateContactSchema,
+  updateContactSchema,
   updateContactFavoriteSchema,
 } from "../schemas/contactsSchemas.js";
 
 export const getAllContacts = async (req, res, next) => {
   try {
-    const contacts = await Contact.find();
+    const { _id: owner } = req.user;
+    const contacts = await Contact.find({ owner });
 
     res.status(200).json(contacts);
   } catch (error) {
@@ -21,7 +22,8 @@ export const getAllContacts = async (req, res, next) => {
 export const getContactById = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const contact = await Contact.findById(id);
+    const { _id: owner } = req.user;
+    const contact = await Contact.findById(id).where("owner").equals(owner);
 
     if (!contact) {
       throw HttpError(404);
@@ -36,7 +38,8 @@ export const getContactById = async (req, res, next) => {
 export const deleteContact = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const deletedContact = await Contact.findByIdAndDelete(id);
+    const { _id: owner } = req.user;
+    const deletedContact = await Contact.findByIdAndDelete(id).where("owner").equals(owner);
 
     if (!deletedContact) {
       throw HttpError(404, "Not found");
@@ -55,7 +58,8 @@ export const createContact = async (req, res, next) => {
       throw HttpError(400, error.message);
     }
 
-    const newContact = await Contact.create(req.body);
+    const { _id: owner } = req.user;
+    const newContact = await Contact.create({ ...req.body, owner });
     res.status(201).json(newContact);
   } catch (error) {
     next(error);
@@ -73,8 +77,11 @@ export const updateContact = async (req, res, next) => {
     }
 
     const { id } = req.params;
-  
-    const updatedContact = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+    const { _id: owner } = req.user;
+    
+    const updatedContact = await Contact.findByIdAndUpdate(id, req.body, {
+      new: true,
+    }).where("owner").equals(owner);
     if (!updatedContact) {
       throw HttpError(404, "Not found");
     }
@@ -86,20 +93,22 @@ export const updateContact = async (req, res, next) => {
 };
 
 export const updateContactFavoriteOption = async (req, res, next) => {
-    try {
-      
-    if (!('favorite' in req.body)) {
-      throw HttpError(400, 'Missing field: favorite');
+  try {
+    if (!("favorite" in req.body)) {
+      throw HttpError(400, "Missing field: favorite");
     }
-   
+
     const { error } = updateContactFavoriteSchema.validate(req.body);
     if (error) {
       throw HttpError(400, error.message);
     }
 
     const { id } = req.params;
-  
-    const updatedContact = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+    const { _id: owner } = req.user;
+
+    const updatedContact = await Contact.findByIdAndUpdate(id, req.body, {
+      new: true,
+    }).where("owner").equals(owner);
     if (!updatedContact) {
       throw HttpError(404, "Not found");
     }
